+++ assets/javascripts/kitten/components/form/password-input/test.js
@@ -1,15 +1,15 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { PasswordInput } from '../../components/form/password-input'
+import { PasswordInput } from './index'
 
 describe('<PasswordInput />', () => {
   it('should match its empty snapshot', () => {
     const tree = renderer
       .create(
         <PasswordInput
-          textInputProps={{}}
           iconLabel="Custom icon label"
           hiddenIconLabel="Custom hidden icon label"
+          id="password"
         />,
       )
       .toJSON()