+++ assets/javascripts/kitten/components/icons/checked-circle-icon.js
@@ -1,9 +1,14 @@
 import React from 'react'
 
-export const CheckedCircleIcon = ({ circleColor, checkedColor, ...props }) => {
+export const CheckedCircleIcon = ({
+  circleColor,
+  checkedColor,
+  title,
+  ...props
+}) => {
   return (
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" {...props}>
-      <title>Checked circle</title>
+      {title && <title>{title}</title>}
       <circle fill={circleColor} cx="10" cy="10" r="10" />
       <path
         fill={checkedColor}
@@ -16,4 +21,5 @@ export const CheckedCircleIcon = ({ circleColor, checkedColor, ...props }) => {
 CheckedCircleIcon.defaultProps = {
   circleColor: '#fff',
   checkedColor: '#333',
+  title: 'Checked circle',
 }