+++ src/components/comments/comment-avatar/index.js
@@ -25,8 +25,6 @@ var _styledComponents = _interopRequireDefault(require("styled-components"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _marger = require("../../../components/layout/marger");
-
 var _buttonImage = require("../../../components/buttons/button-image");
 
 var _text = require("../../../components/typography/text");
@@ -40,12 +38,22 @@ var StyledAvatar = _styledComponents.default.div.withConfig({
   componentId: "sc-10x87zp-0"
 })(["display:flex;flex-direction:column;align-items:center;"]);
 
+var StyledPosition = _styledComponents.default.div.withConfig({
+  displayName: "comment-avatar__StyledPosition",
+  componentId: "sc-10x87zp-1"
+})(["position:relative;margin-bottom:", ";"], (0, _typography.pxToRem)(10));
+
 var buttonImgHuge = (0, _typography.pxToRem)(80);
 var buttonImgBig = (0, _typography.pxToRem)(50);
 var StyledButtonImage = (0, _styledComponents.default)(_buttonImage.ButtonImage).withConfig({
   displayName: "comment-avatar__StyledButtonImage",
-  componentId: "sc-10x87zp-1"
+  componentId: "sc-10x87zp-2"
 })(["border-radius:50%;height:", ";width:", ";@media (min-width:", "px){height:", ";width:", ";}"], buttonImgBig, buttonImgBig, _screenConfig.ScreenConfig.S.min, buttonImgHuge, buttonImgHuge);
+
+var StyledAvatarBadge = _styledComponents.default.div.withConfig({
+  displayName: "comment-avatar__StyledAvatarBadge",
+  componentId: "sc-10x87zp-3"
+})(["position:absolute;right:0;top:0;"]);
 
 var CommentAvatar =
 /*#__PURE__*/
@@ -61,13 +69,18 @@ function (_PureComponent) {
     key: "render",
     value: function render() {
       var avatarImgProps = this.props.avatarImgProps;
-      return _react.default.createElement(StyledAvatar, null, _react.default.createElement(_marger.Marger, {
-        bottom: "1"
-      }, _react.default.createElement(StyledButtonImage, {
+      return _react.default.createElement(StyledAvatar, null, _react.default.createElement(StyledPosition, null, this.renderBadge(), _react.default.createElement(StyledButtonImage, {
         tag: "span",
         withoutPointerEvents: true,
         img: avatarImgProps
       })), this.renderDate());
+    }
+  }, {
+    key: "renderBadge",
+    value: function renderBadge() {
+      var avatarBadge = this.props.avatarBadge;
+      if (!avatarBadge) return;
+      return _react.default.createElement(StyledAvatarBadge, null, avatarBadge);
     }
   }, {
     key: "renderDate",
@@ -87,8 +100,10 @@ function (_PureComponent) {
 exports.CommentAvatar = CommentAvatar;
 CommentAvatar.propTypes = {
   avatarImgProps: _propTypes.default.object.isRequired,
-  commentDate: _propTypes.default.string
+  commentDate: _propTypes.default.string,
+  avatarBadge: _propTypes.default.node
 };
 CommentAvatar.defaultProps = {
-  commentDate: ''
+  commentDate: '',
+  avatarBadge: ''
 };