+++ src/components/comments/comment/index.js
@@ -9,21 +9,13 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports.Comment = void 0;
 
+var _slicedToArray2 = _interopRequireDefault(require("@babel/runtime/helpers/slicedToArray"));
+
 var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
-
-var _classCallCheck2 = _interopRequireDefault(require("@babel/runtime/helpers/classCallCheck"));
-
-var _createClass2 = _interopRequireDefault(require("@babel/runtime/helpers/createClass"));
-
-var _possibleConstructorReturn2 = _interopRequireDefault(require("@babel/runtime/helpers/possibleConstructorReturn"));
-
-var _getPrototypeOf2 = _interopRequireDefault(require("@babel/runtime/helpers/getPrototypeOf"));
-
-var _inherits2 = _interopRequireDefault(require("@babel/runtime/helpers/inherits"));
 
 var _react = _interopRequireWildcard(require("react"));
 
-var _styledComponents = _interopRequireDefault(require("styled-components"));
+var _styledComponents = _interopRequireWildcard(require("styled-components"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
@@ -39,6 +31,12 @@ var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-co
 
 var _typography = require("../../../helpers/utils/typography");
 
+var _likeButton = require("./components/like-button");
+
+var _elementHelper = _interopRequireDefault(require("../../../helpers/dom/element-helper"));
+
+var desktopPadding = (0, _typography.pxToRem)(30);
+var tabletAndMobilePadding = (0, _typography.pxToRem)(20);
 var StyledContentText = (0, _styledComponents.default)(_text.Text).withConfig({
   displayName: "comment__StyledContentText",
   componentId: "sc-8s8e85-0"
@@ -47,7 +45,7 @@ var StyledContentText = (0, _styledComponents.default)(_text.Text).withConfig({
 var StyledCommentArrow = _styledComponents.default.span.withConfig({
   displayName: "comment__StyledCommentArrow",
   componentId: "sc-8s8e85-1"
-})(["position:absolute;top:", ";display:block;width:0;height:0;border-width:", ";border-style:solid;border-color:transparent;border-right-color:", ";left:-", ";@media (min-width:", "px){top:", ";}"], (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(10), _colorsConfig.default.background3, (0, _typography.pxToRem)(20), _screenConfig.ScreenConfig.S.min, (0, _typography.pxToRem)(35));
+})(["position:absolute;top:", ";display:block;width:0;height:0;border-width:", ";border-style:solid;border-color:transparent;border-right-color:", ";left:", ";@media (min-width:", "px){top:", ";}"], (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(10), _colorsConfig.default.background3, (0, _typography.pxToRem)(-20), _screenConfig.ScreenConfig.S.min, (0, _typography.pxToRem)(35));
 
 var StyledGrid = _styledComponents.default.div.withConfig({
   displayName: "comment__StyledGrid",
@@ -57,64 +55,87 @@ var StyledGrid = _styledComponents.default.div.withConfig({
 var StyledCommentContainer = _styledComponents.default.span.withConfig({
   displayName: "comment__StyledCommentContainer",
   componentId: "sc-8s8e85-3"
-})(["position:relative;margin-left:", ";@media (min-width:", "px){margin-left:", ",},"], (0, _typography.pxToRem)(20), _screenConfig.ScreenConfig.S.min, (0, _typography.pxToRem)(35));
+})(["position:relative;margin-left:", ";flex:1;@media (min-width:", "px){margin-left:", ";}"], (0, _typography.pxToRem)(20), _screenConfig.ScreenConfig.S.min, (0, _typography.pxToRem)(35));
 
 var StyledCommentContent = _styledComponents.default.div.withConfig({
   displayName: "comment__StyledCommentContent",
   componentId: "sc-8s8e85-4"
-})(["border-width:", ";background-color:", ";border-color:", ";color:", ";padding:", ";font-size:", ";"], (0, _typography.pxToRem)(2), _colorsConfig.default.background3, _colorsConfig.default.background3, _colorsConfig.default.font1, (0, _typography.pxToRem)(30), (0, _typography.stepToRem)(0));
+})(["border-width:", ";background-color:", ";border-color:", ";color:", ";font-size:", ";padding:", ";@media (min-width:", "px){padding-top:", ";padding-right:", ";padding-left:", ";}"], (0, _typography.pxToRem)(2), _colorsConfig.default.background3, _colorsConfig.default.background3, _colorsConfig.default.font1, (0, _typography.stepToRem)(0), tabletAndMobilePadding, _screenConfig.ScreenConfig.M.min, desktopPadding, desktopPadding, desktopPadding);
+
+var StyledMargerText = _styledComponents.default.div.withConfig({
+  displayName: "comment__StyledMargerText",
+  componentId: "sc-8s8e85-5"
+})(["", " ", ""], function (_ref) {
+  var hasLikeButton = _ref.hasLikeButton;
+  return hasLikeButton && (0, _styledComponents.css)(["margin-bottom:", ";@media (min-width:", "px){margin-bottom:", ";}"], (0, _typography.pxToRem)(15), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(20));
+}, function (_ref2) {
+  var ownerName = _ref2.ownerName;
+  return ownerName && (0, _styledComponents.css)(["margin-top:", ";"], (0, _typography.pxToRem)(10));
+});
 
 var StyledBottomNotes = (0, _styledComponents.default)(_text.Text).withConfig({
   displayName: "comment__StyledBottomNotes",
-  componentId: "sc-8s8e85-5"
+  componentId: "sc-8s8e85-6"
 })(["padding-left:", ";margin-top:0;"], (0, _typography.pxToRem)(30));
 
-var Comment =
-/*#__PURE__*/
-function (_PureComponent) {
-  (0, _inherits2.default)(Comment, _PureComponent);
+var StyledLikeButtonBox = _styledComponents.default.div.withConfig({
+  displayName: "comment__StyledLikeButtonBox",
+  componentId: "sc-8s8e85-7"
+})(["position:absolute;right:0;margin-top:", ";margin-right:", ";@media (min-width:", "px){margin-right:", ";}"], (0, _typography.pxToRem)(-20), (0, _typography.pxToRem)(20), _screenConfig.ScreenConfig.M.min, (0, _typography.pxToRem)(30));
 
-  function Comment() {
-    (0, _classCallCheck2.default)(this, Comment);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(Comment).apply(this, arguments));
-  }
+var Comment = function Comment(_ref3) {
+  var text = _ref3.text,
+      ownerName = _ref3.ownerName,
+      avatarImgProps = _ref3.avatarImgProps,
+      commentDate = _ref3.commentDate,
+      bottomNotes = _ref3.bottomNotes,
+      likeButtonProps = _ref3.likeButtonProps,
+      avatarBadge = _ref3.avatarBadge,
+      props = (0, _objectWithoutProperties2.default)(_ref3, ["text", "ownerName", "avatarImgProps", "commentDate", "bottomNotes", "likeButtonProps", "avatarBadge"]);
+  var likeButtonElement = (0, _react.useRef)(null);
 
-  (0, _createClass2.default)(Comment, [{
-    key: "render",
-    value: function render() {
-      var _this$props = this.props,
-          text = _this$props.text,
-          ownerName = _this$props.ownerName,
-          avatarImgProps = _this$props.avatarImgProps,
-          commentDate = _this$props.commentDate,
-          bottomNotes = _this$props.bottomNotes,
-          props = (0, _objectWithoutProperties2.default)(_this$props, ["text", "ownerName", "avatarImgProps", "commentDate", "bottomNotes"]);
-      return _react.default.createElement(StyledGrid, null, _react.default.createElement(_commentAvatar.CommentAvatar, {
-        avatarImgProps: avatarImgProps,
-        commentDate: commentDate
-      }), _react.default.createElement(StyledCommentContainer, null, _react.default.createElement(StyledCommentContent, null, ownerName && _react.default.createElement(_marger.Marger, {
-        bottom: "1"
-      }, _react.default.createElement(_text.Text, {
-        color: "font1",
-        size: "tiny",
-        weight: "regular"
-      }, ownerName)), _react.default.createElement(_marger.Marger, {
-        top: ownerName ? 1 : null
-      }, _react.default.createElement(StyledContentText, {
-        color: "font1",
-        weight: "light"
-      }, text)), _react.default.createElement(StyledCommentArrow, null)), bottomNotes && _react.default.createElement(_marger.Marger, {
-        top: ".5"
-      }, _react.default.createElement(StyledBottomNotes, {
-        tag: "p",
-        color: "font1",
-        size: "micro",
-        weight: "bold"
-      }, bottomNotes))));
+  var _useState = (0, _react.useState)(0),
+      _useState2 = (0, _slicedToArray2.default)(_useState, 2),
+      likeButtonWidth = _useState2[0],
+      setLikeButtonWidth = _useState2[1];
+
+  (0, _react.useEffect)(function () {
+    var marginLeftAndRight = 40;
+
+    var elementWidth = _elementHelper.default.getComputedWidth(likeButtonElement.current);
+
+    setLikeButtonWidth(elementWidth + marginLeftAndRight);
+  }, []);
+  return _react.default.createElement(StyledGrid, null, _react.default.createElement(_commentAvatar.CommentAvatar, {
+    avatarImgProps: avatarImgProps,
+    commentDate: commentDate,
+    avatarBadge: avatarBadge
+  }), _react.default.createElement(StyledCommentContainer, null, _react.default.createElement(StyledCommentContent, null, ownerName && _react.default.createElement(_marger.Marger, {
+    bottom: "1"
+  }, _react.default.createElement(_text.Text, {
+    color: "font1",
+    size: "tiny",
+    weight: "regular"
+  }, ownerName)), _react.default.createElement(StyledMargerText, {
+    hasLikeButton: !!likeButtonProps.children,
+    ownerName: ownerName
+  }, _react.default.createElement(StyledContentText, {
+    color: "font1",
+    weight: "light"
+  }, text)), _react.default.createElement(StyledCommentArrow, null)), likeButtonProps.children && _react.default.createElement(StyledLikeButtonBox, {
+    ref: likeButtonElement
+  }, _react.default.createElement(_likeButton.LikeButton, likeButtonProps)), bottomNotes && _react.default.createElement(_marger.Marger, {
+    top: ".5",
+    style: {
+      marginRight: likeButtonWidth
     }
-  }]);
-  return Comment;
-}(_react.PureComponent);
+  }, _react.default.createElement(StyledBottomNotes, {
+    tag: "div",
+    color: "font1",
+    size: "micro",
+    weight: "bold"
+  }, bottomNotes))));
+};
 
 exports.Comment = Comment;
 Comment.propTypes = {
@@ -122,9 +143,13 @@ Comment.propTypes = {
   ownerName: _propTypes.default.string,
   avatarImgProps: _propTypes.default.object.isRequired,
   commentDate: _propTypes.default.string.isRequired,
-  bottomNotes: _propTypes.default.node
+  bottomNotes: _propTypes.default.node,
+  likeButtonProps: _propTypes.default.object,
+  avatarBadge: _propTypes.default.node
 };
 Comment.defaultProps = {
   bottomNotes: '',
-  ownerName: ''
+  ownerName: '',
+  likeButtonProps: {},
+  avatarBadge: ''
 };