+++ src/components/icons/bubble-icon/index.js
@@ -5,7 +5,7 @@ var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefau
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.PasswordIcon = void 0;
+exports.BubbleIcon = void 0;
 
 var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
 
@@ -13,26 +13,29 @@ var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/h
 
 var _react = _interopRequireDefault(require("react"));
 
-var PasswordIcon = function PasswordIcon(_ref) {
+var _propTypes = _interopRequireDefault(require("prop-types"));
+
+var BubbleIcon = function BubbleIcon(_ref) {
   var color = _ref.color,
       title = _ref.title,
       props = (0, _objectWithoutProperties2.default)(_ref, ["color", "title"]);
   return _react.default.createElement("svg", (0, _extends2.default)({
+    width: "14",
+    height: "12",
     xmlns: "http://www.w3.org/2000/svg",
-    viewBox: "0 0 18 12",
-    fill: color,
-    width: "18"
+    fill: color
   }, props), title && _react.default.createElement("title", null, title), _react.default.createElement("path", {
-    d: "M9 0C4 0 0 4 0 6s4 6 9 6 9-4 9-6-4-6-9-6zm0 10c-2.2 0-4-1.8-4-4s1.8-4 4-4 4 1.8 4 4-1.8 4-4 4z"
-  }), _react.default.createElement("circle", {
-    cx: "9",
-    cy: "6",
-    r: "2"
+    d: "M3.417 11.155L1.5 12.732l.431-2.594C.735 9.062 0 7.604 0 6c0-3.314 3.134-6 7-6s7 2.686 7 6-3.134 6-7 6a7.865 7.865 0 0 1-3.583-.845z",
+    fillRule: "evenodd"
   }));
 };
 
-exports.PasswordIcon = PasswordIcon;
-PasswordIcon.defaultProps = {
+exports.BubbleIcon = BubbleIcon;
+BubbleIcon.propTypes = {
+  color: _propTypes.default.string,
+  title: _propTypes.default.string
+};
+BubbleIcon.defaultProps = {
   color: '#222',
   title: ''
 };